--- a/Modules/Core/Common/itk-module.cmake
+++ b/Modules/Core/Common/itk-module.cmake
@@ -13,6 +13,7 @@ endif()
 itk_module(ITKCommon
   ENABLE_SHARED
   DEPENDS
+    ITKEigen3
     ITKKWIML
     ${ITKCOMMON_TBB_DEPENDS}
   PRIVATE_DEPENDS