--- a/Modules/IO/BMP/itk-module.cmake
+++ b/Modules/IO/BMP/itk-module.cmake
@@ -3,7 +3,7 @@ Microsoft Bitmap File Format (BMP).")
 
 itk_module(ITKIOBMP
   ENABLE_SHARED
-  PRIVATE_DEPENDS
+  DEPENDS
     ITKIOImageBase
   TEST_DEPENDS
     ITKTestKernel