--- a/Modules/IO/Bruker/itk-module.cmake
+++ b/Modules/IO/Bruker/itk-module.cmake
@@ -2,7 +2,7 @@ set(DOCUMENTATION "This module contains classes that read Bruker image files.")
 
 itk_module(ITKIOBruker
   ENABLE_SHARED
-  PRIVATE_DEPENDS
+  DEPENDS
     ITKIOImageBase
   TEST_DEPENDS
     ITKTestKernel