--- a/Modules/IO/MRC/itk-module.cmake
+++ b/Modules/IO/MRC/itk-module.cmake
@@ -3,7 +3,7 @@ files in the MRC file format. https://en.wikipedia.org/wiki/MRC_(file_format) ")
 
 itk_module(ITKIOMRC
   ENABLE_SHARED
-  PRIVATE_DEPENDS
+  DEPENDS
     ITKIOImageBase
   TEST_DEPENDS
     ITKTestKernel