--- a/Modules/IO/Meta/itk-module.cmake
+++ b/Modules/IO/Meta/itk-module.cmake
@@ -8,7 +8,6 @@ itk_module(ITKIOMeta
   ENABLE_SHARED
   DEPENDS
     ITKMetaIO
-  PRIVATE_DEPENDS
     ITKIOImageBase
   TEST_DEPENDS
     ITKTestKernel