--- a/Modules/IO/NRRD/itk-module.cmake
+++ b/Modules/IO/NRRD/itk-module.cmake
@@ -4,8 +4,9 @@ set(DOCUMENTATION "This module contains an ImageIO class to read and write the
 
 itk_module(ITKIONRRD
   ENABLE_SHARED
-  PRIVATE_DEPENDS
+  DEPENDS
     ITKIOImageBase
+  PRIVATE_DEPENDS
     ITKNrrdIO
   TEST_DEPENDS
     ITKTestKernel