--- a/Modules/IO/SpatialObjects/itk-module.cmake
+++ b/Modules/IO/SpatialObjects/itk-module.cmake
@@ -6,8 +6,9 @@ format.")
 
 
 itk_module(ITKIOSpatialObjects
-  PRIVATE_DEPENDS
+  DEPENDS
     ITKSpatialObjects
+  PRIVATE_DEPENDS
     ITKIOXML
     ITKMesh
   TEST_DEPENDS