--- a/Modules/Remote/SmoothingRecursiveYvvGaussianFilter.remote.cmake
+++ b/Modules/Remote/SmoothingRecursiveYvvGaussianFilter.remote.cmake
@@ -2,6 +2,6 @@
 itk_fetch_module(SmoothingRecursiveYvvGaussianFilter
   "GPU and CPU Young & Van Vliet Recursive Gaussian Smoothing Filter: https://hdl.handle.net/10380/3425"
   #UPSTREAM_REPO GIT_REPOSITORY ${git_protocol}://github.com/Inria-Asclepios/SmoothingRecursiveYvvGaussianFilter
-  GIT_REPOSITORY ${git_protocol}://github.com/InsightSoftwareConsortium/SmoothingRecursiveYvvGaussianFilter.git
-  GIT_TAG 978da8a4a4c1d3739462f5607937119423a72825
+  GIT_REPOSITORY ${git_protocol}://github.com/InsightSoftwareConsortium/ITKSmoothingRecursiveYvvGaussianFilter.git
+  GIT_TAG 2b14aa9167d104cce5672b0c9bcbd2b72a4d741a
   )