--- a/Modules/ThirdParty/Eigen3/src/itk_eigen.h.in
+++ b/Modules/ThirdParty/Eigen3/src/itk_eigen.h.in
@@ -1,0 +1,35 @@
+/*=========================================================================
+ *
+ *  Copyright Insight Software Consortium
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0.txt
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *=========================================================================*/
+
+#ifndef itk_eigen_h
+#define itk_eigen_h
+
+// ITK should only be using MPL2 Eigen code.
+#ifndef EIGEN_MPL2_ONLY
+#define EIGEN_MPL2_ONLY
+#endif
+
+/* Use the eigen library configured for ITK.  */
+#cmakedefine ITK_USE_SYSTEM_EIGEN
+#ifdef ITK_USE_SYSTEM_EIGEN
+# define ITK_EIGEN(x) <Eigen/x>
+#else
+# define ITK_EIGEN(x) <itkeigen/Eigen/x>
+#endif
+
+#endif