--- a/Modules/ThirdParty/Eigen3/src/itkeigen/.gitattributes
+++ b/Modules/ThirdParty/Eigen3/src/itkeigen/.gitattributes
@@ -1,0 +1,3 @@
+* -whitespace
+Eigen/src/misc/lapacke.h hooks-max-size=1500000
+Eigen/src/misc/lapacke.h hooks.MaxObjectKiB=2048