--- a/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Dense
+++ b/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Dense
@@ -1,0 +1,7 @@
+#include "Core"
+#include "LU"
+#include "Cholesky"
+#include "QR"
+#include "SVD"
+#include "Geometry"
+#include "Eigenvalues"