--- a/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Eigen
+++ b/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Eigen
@@ -1,0 +1,2 @@
+#include "Dense"
+#include "Sparse"