--- a/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/src/Core/util/NonMPL2.h
+++ b/Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/src/Core/util/NonMPL2.h
@@ -1,0 +1,3 @@
+#ifdef EIGEN_MPL2_ONLY
+#error Including non-MPL2 code in EIGEN_MPL2_ONLY mode
+#endif