--- a/Modules/ThirdParty/Eigen3/src/itkeigen/README.md
+++ b/Modules/ThirdParty/Eigen3/src/itkeigen/README.md
@@ -1,0 +1,3 @@
+**Eigen is a C++ template library for linear algebra: matrices, vectors, numerical solvers, and related algorithms.**
+
+For more information go to http://eigen.tuxfamily.org/.