… Modules/Core/Common/include/itkSymmetricEigenAnalysis.h
… Modules/Core/Common/test/itkSymmetricEigenAnalysisTest.cxx
… Modules/Filtering/ImageIntensity/test/CMakeLists.txt
… Modules/Filtering/ImageIntensity/test/itkSymmetricEigenAnalysisImageFilterTest.cxx
… Modules/ThirdParty/Eigen3/CMakeLists.txt
… Modules/ThirdParty/Eigen3/src/CMakeLists.txt
… Modules/ThirdParty/Eigen3/src/itkeigen/CMakeLists.txt
… Modules/ThirdParty/Eigen3/src/itkeigen/COPYING.MPL2
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/CMakeLists.txt
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/CholmodSupport
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Core
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Eigenvalues
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Geometry
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/IterativeLinearSolvers
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Jacobi
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/LU
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/MetisSupport
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/OrderingMethods
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/PaStiXSupport
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/PardisoSupport
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/QR
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/QtAlignedMalloc
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SPQRSupport
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SVD
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/Sparse
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SparseCholesky
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SparseCore
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SparseLU
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/SparseQR
… Modules/ThirdParty/Eigen3/src/itkeigen/Eigen/StdDeque