The idea is to introduce Eigen3 into ITK.
Please note that ITKEigen is already a module.
The name of this module is ITKEigen3

It uses a modified Eigen3 stored in:
https://gitlab.kitware.com/third-party/eigen

I forked and modified the existing branch `for/vtk`
into `for/itk`.

-- Note that for development I am using my own fork:
https://gitlab.kitware.com/phcerdan/eigen/tree/for/itk --

A MR is in process requesting to create the branch for/itk
https://gitlab.kitware.com/third-party/eigen/merge_requests/1

'for/itk' differs from 'for/vtk' that it allows the developer
to use find_package(Eigen3) in the internal itkeigen, just
populating the Eigen3_DIR with the internal folder:

```cmake
set(_internal_cmake_eigen3)
list(GET ITKEigen3_INCLUDE_DIRS 0 _internal_cmake_eigen3)
set(Eigen3_DIR "${_internal_cmake_eigen3}/itkeigen")
```
See https://github.com/phcerdan/ITKTotalVariation for up-to-date
handling of Eigen3_DIR.

Eigen3 is faster than non-optimized implementations of
lapack. VNL has the netlib (non-optimized) CLAPACK implementation.
CLAPACK is the c-version of the netlib lapack libraries (written in fortran),
it used the f2c conversor from fortran to c. It is not maintained anymore.

One of the purposes on introduced Eigen3 is to speed up eigen
decomposition. Also having the most used linear algebra library
in c++ opens opportunities for bridging with other libraries.

The first application for the module is to use it in a external remote
module TotalVariation, that performs total variation regularization for
denoising methods:
https://github.com/phcerdan/ITKTotalVariation
The module is a wrap around the third party library:
Upstream: https://github.com/albarji/proxTV
Used fork: https://github.com/phcerdan/proxTV/tree/use_eigen
that requires Eigen3.