@thewtex 2018-11-14
> And so gradual transition from VXL to Eigen begins

:sparkles: :sparkling_heart: !

Wonderful!

@phcerdan could you please use the Eigen somewhere in ITK so our CI tests are exercising its functionality in some way?