@phcerdan 2018-11-15
Last commit adds internal usage for computing eigenvectors in symmetric matrices:
 Check commit message from: https://github.com/InsightSoftwareConsortium/ITK/pull/183/commits/dc8fb1827605a2680e4433ddb537cfb6e2b0a41f

WIP: Use Eigen3 internally for EigenAnalysis

Implement it using Eigen3, running the tests shows already a performance
increase.
Flag m_UseEigenLibrary set to True for checking in CI

Issues: There are billions of -Wshadow warnings in gcc.

Question/TODO: How do I use the target Eigen3::Eigen
internally in ITKCommon instead of just including the header.
ITKEigen3.cmake does not have any targets file,
and there is no library.
It would be awesome to set the target_include_directories(eigen) as SYSTEM, to
avoid the warnings.

A non-scientific tiny perf comparison using tests:
`ctest -R "Symm"`
Comparing ITKCommon:
old: 0.11
new: 0.08

const-correct netlib (OLD):
```

      Start 1456: itkSymmetricEigenAnalysisImageFilterTestOrderByValue
 8/12 Test #1456: itkSymmetricEigenAnalysisImageFilterTestOrderByValue .......   Passed    0.03 sec
      Start 1457: itkSymmetricEigenAnalysisImageFilterTestOrderByMagnitude
 9/12 Test #1457: itkSymmetricEigenAnalysisImageFilterTestOrderByMagnitude ...   Passed    0.02 sec
      Start 1458: itkSymmetricEigenAnalysisImageFilterTestDoNotOrder
10/12 Test #1458: itkSymmetricEigenAnalysisImageFilterTestDoNotOrder .........   Passed    0.02 sec
      Start 2031: itkSymmetricForcesDemonsRegistrationFilterTest
11/12 Test #2031: itkSymmetricForcesDemonsRegistrationFilterTest .............   Passed    0.17 sec
      Start 2044: itkFastSymmetricForcesDemonsRegistrationFilterTest
12/12 Test #2044: itkFastSymmetricForcesDemonsRegistrationFilterTest .........   Passed    0.11 sec

100% tests passed, 0 tests failed out of 12

Label Time Summary:
ITKCommon                       =   0.11 sec*proc (5 tests)
```

Eigen3 (NEW):
```
      Start 1456: itkSymmetricEigenAnalysisImageFilterTestOrderByValue
 8/12 Test #1456: itkSymmetricEigenAnalysisImageFilterTestOrderByValue .......   Passed    0.01 sec
      Start 1457: itkSymmetricEigenAnalysisImageFilterTestOrderByMagnitude
 9/12 Test #1457: itkSymmetricEigenAnalysisImageFilterTestOrderByMagnitude ...   Passed    0.01 sec
      Start 1458: itkSymmetricEigenAnalysisImageFilterTestDoNotOrder
10/12 Test #1458: itkSymmetricEigenAnalysisImageFilterTestDoNotOrder .........   Passed    0.01 sec
      Start 2031: itkSymmetricForcesDemonsRegistrationFilterTest
11/12 Test #2031: itkSymmetricForcesDemonsRegistrationFilterTest .............   Passed    0.14 sec
      Start 2044: itkFastSymmetricForcesDemonsRegistrationFilterTest
12/12 Test #2044: itkFastSymmetricForcesDemonsRegistrationFilterTest .........   Passed    0.11 sec

100% tests passed, 0 tests failed out of 12

Label Time Summary:
ITKCommon                       =   0.08 sec*proc (5 tests)
```