@thewtex 2018-11-15
> Question/TODO: How do I use the target Eigen3::Eigen
> internally in ITKCommon instead of just including the header.
> ITKEigen3.cmake does not have any targets file,
> and there is no library.
> It would be awesome to set the target_include_directories(eigen) as SYSTEM, to
> avoid the warnings.

An `INTERFACE` library is what is desired here?:

https://cmake.org/cmake/help/latest/command/add_library.html#id6
http://mariobadr.com/creating-a-header-only-library-with-cmake.html