@phcerdan 2018-11-27
Failure on CI seems related to missing data:
```
2018-11-27T20:00:00.5885140Z FAILED: ExternalData/Modules/IO/TIFF/test/Input/3d_spacing.tif-hash-stamp ExternalData/Modules/IO/TIFF/test/Input/3d_spacing.tif 
2018-11-27T20:00:00.5886720Z cd /Users/vsts/agent/2.142.1/work/1/s-build && /usr/local/Cellar/cmake/3.12.3/bin/cmake -Drelative_top=/Users/vsts/agent/2.142.1/work/1/s-build -Dfile=/Users/vsts/agent/2.142.1/work/1/s-build/ExternalData/Modules/IO/TIFF/test/Input/3d_spacing.tif -Dname=/Users/vsts/agent/2.142.1/work/1/s/Modules/IO/TIFF/test/Input/3d_spacing.tif -Dexts=.md5 -DExternalData_ACTION=fetch -DExternalData_CONFIG=/Users/vsts/agent/2.142.1/work/1/s-build/ITKData_config.cmake -P /usr/local/Cellar/cmake/3.12.3/share/cmake/Modules/ExternalData.cmake
2018-11-27T20:00:00.5887880Z -- Fetching "https://insightsoftwareconsortium.github.io/ITKTestingData/MD5/f81aa486be7bd1d51fdcac68000e08a2"
2018-11-27T20:00:00.5888960Z -- Fetching "https://data.kitware.com:443/api/v1/file/hashsum/MD5/f81aa486be7bd1d51fdcac68000e08a2/download"
2018-11-27T20:00:00.5975840Z -- Fetching "https://itk.org/files/ExternalData/MD5/f81aa486be7bd1d51fdcac68000e08a2"
2018-11-27T20:00:00.5977290Z -- Fetching "https://slicer.kitware.com/midas3/api/rest?method=midas.bitstream.download&checksum=f81aa486be7bd1d51fdcac68000e08a2&algorithm=MD5"
2018-11-27T20:00:00.5978330Z -- [download 0% complete]
2018-11-27T20:00:00.5979360Z -- [download 100% complete]
2018-11-27T20:00:00.5979900Z CMake Error at /usr/local/Cellar/cmake/3.12.3/share/cmake/Modules/ExternalData.cmake:1121 (message):
2018-11-27T20:00:00.5980360Z   
2018-11-27T20:00:00.5980810Z 
2018-11-27T20:00:00.5981290Z   Object MD5=f81aa486be7bd1d51fdcac68000e08a2 not found at:
2018-11-27T20:00:00.6081800Z 
2018-11-27T20:00:00.6082270Z     https://insightsoftwareconsortium.github.io/ITKTestingData/MD5/f81aa486be7bd1d51fdcac68000e08a2 ("HTTP response code said error")
2018-11-27T20:00:00.6082360Z     https://data.kitware.com:443/api/v1/file/hashsum/MD5/f81aa486be7bd1d51fdcac68000e08a2/download ("HTTP response code said error")
2018-11-27T20:00:00.6083620Z     https://itk.org/files/ExternalData/MD5/f81aa486be7bd1d51fdcac68000e08a2 ("Couldn't connect to server")
2018-11-27T20:00:00.6084030Z     https://slicer.kitware.com/midas3/api/rest?method=midas.bitstream.download&checksum=f81aa486be7bd1d51fdcac68000e08a2&algorithm=MD5 (wrong hash MD5=980d29c660124c77d48524bcb9ea384c)
2018-11-27T20:00:00.6084120Z 
```