@jcfr 2018-11-27
> Is that update going on right now?

After double checking, the cluster update is planned for tomorrow morning.  This means the issue is something else.