@thewtex 2018-11-27
data.kitware.com has been occasionally failing to serve the files.

Issue #159 will improve reliablity.

In the meantime, @phcerdan please try `git commit --amend` followed by `git review-push --force`.